+++ src/core/algorithms/md/hymd/preprocessing/similarity_measure/smith_waterman_gotoh.h
@@ -3,6 +3,6 @@
 #include <string>
 
 namespace algos::hymd::preprocessing::similarity_measure {
-float NormalizedSmithWatermanGotoh(std::string const& a, std::string const& b,
-                                   float gap_value = -0.5f);
+double NormalizedSmithWatermanGotoh(std::string const& a, std::string const& b,
+                                    double gap_value = -0.5);
 }  // namespace algos::hymd::preprocessing::similarity_measure
